[Settings] Separate "Applications" and "Removable media"
